fabric/mr: Use void * for local descriptor in place of uint64_t
The size of the local descriptor may vary.  Use void * instead
of uint64_t.  Providers may still encode keys directly into
the pointer, provided that they fit.

+++ prov/psm/src/psmx_mr.c
@@ -183,7 +183,7 @@ static int psmx_mr_reg(fid_t fid, const void *buf, size_t len,
 	fid_mr->mr.fid.fclass = FID_CLASS_MR;
 	fid_mr->mr.fid.context = context;
 	fid_mr->mr.fid.ops = &psmx_fi_ops;
-	fid_mr->mr.mem_desc = (uint64_t)(uintptr_t)fid_mr;
+	fid_mr->mr.mem_desc = fid_mr;
 	fid_mr->mr.key = (uint64_t)(uintptr_t)fid_mr; /* requested_key is ignored */
 	fid_mr->domain = fid_domain;
 	fid_mr->signature = PSMX_MR_SIGNATURE;
@@ -221,7 +221,7 @@ static int psmx_mr_regv(fid_t fid, const struct iovec *iov, size_t count,
 	fid_mr->mr.fid.fclass = FID_CLASS_MR;
 	fid_mr->mr.fid.context = context;
 	fid_mr->mr.fid.ops = &psmx_fi_ops;
-	fid_mr->mr.mem_desc = (uint64_t)(uintptr_t)fid_mr;
+	fid_mr->mr.mem_desc = fid_mr;
 	fid_mr->mr.key = (uint64_t)(uintptr_t)fid_mr; /* requested_key is ignored */
 	fid_mr->domain = fid_domain;
 	fid_mr->signature = PSMX_MR_SIGNATURE;
@@ -265,7 +265,7 @@ static int psmx_mr_regattr(fid_t fid, const struct fi_mr_attr *attr,
 	fid_mr->mr.fid.size = sizeof(struct fid_mr);
 	fid_mr->mr.fid.fclass = FID_CLASS_MR;
 	fid_mr->mr.fid.ops = &psmx_fi_ops;
-	fid_mr->mr.mem_desc = (uint64_t)(uintptr_t)fid_mr;
+	fid_mr->mr.mem_desc = fid_mr;
 	fid_mr->mr.key = (uint64_t)(uintptr_t)fid_mr;
 	fid_mr->domain = fid_domain;
 	fid_mr->signature = PSMX_MR_SIGNATURE;
